Move to common to be able to reuse it
