+++ nucliadb/src/nucliadb/writer/api/v1/resource.py
@@ -28,6 +28,7 @@ from fastapi_versioning import version
 from starlette.requests import Request
 
 from nucliadb.common import datamanagers
+from nucliadb.common.back_pressure import maybe_back_pressure
 from nucliadb.common.context.fastapi import get_app_context
 from nucliadb.common.maindb.driver import Driver
 from nucliadb.common.maindb.exceptions import ConflictError, NotFoundError
@@ -45,7 +46,6 @@ from nucliadb.writer.api.v1.router import (
     api,
 )
 from nucliadb.writer.api.v1.slug import ensure_slug_uniqueness, noop_context_manager
-from nucliadb.writer.back_pressure import maybe_back_pressure
 from nucliadb.writer.resource.audit import parse_audit
 from nucliadb.writer.resource.basic import (
     parse_basic_creation,
